refactor: 単一コンポーネントの命名の見直し #124
