+++ app/javascript/app.vue
@@ -10,7 +10,7 @@
 </template>
 
 <script>
-import TheHeader from "components/layout/TheHeader"
+import TheHeader from "components/shared/TheHeader"
 
 export default {
   components: {
