+++ app/javascript/components/shared/TheMessage.vue
@@ -10,7 +10,7 @@
 </template>
 <script>
 export default {
-  name: "Message",
+  name: "TheMessage",
   props: {
     alert: {
       type: String,
